Provide flexible image selection
Commit introducess that ability to select which image is used for the
provisioning of virtual instances, also sets default to AlmaLinux 8 in
the process. Either a full path to a custom image can be provided or a
Marketplace URN. When a URN in provided, the image_plan parameter also
must be set. If the Marketplace offering does not have an associated
plan then set to null.

Image_plan values are in a psuedo URN format, name:product:publisher.
The values should be obtained using the "az vm image show" command.